fix: centralize rpc shared helpers to avoid bare import leaks
