--- packages/sdk/server/rpc/delegate-utils.ts
@@ -1,6 +1,0 @@
-let commandCounter = 0;
-
-export function getNextCommandId(): number {
-  commandCounter = (commandCounter + 1) % Number.MAX_SAFE_INTEGER;
-  return commandCounter;
-}
